After extensive wrestling with lifetime parameters, the new context finally goes through.

+++ source/air/src/ast.rs
@@ -8,10 +8,10 @@ pub struct Span {
     pub raw_span: RawSpan,
     pub as_string: String, // if we can't print (description, raw_span), print as_string instead
     pub filename: String,
-    pub start_row: u32,
-    pub start_col: u32,
-    pub end_row: u32,
-    pub end_col: u32,
+    pub start_row: usize,
+    pub start_col: usize,
+    pub end_row: usize,
+    pub end_col: usize,
 }
 pub type SpanOption = Rc<Option<Span>>;
